-- Multiple same name positions problem, solved!! --
	The new positions with retain the original name
	given to them. The position that used to have
	the original name will be given the same name but
	with its creation time-stamp appended to the end.
	The orignal name will be kept in...
		appProperties: actual_name

Notes:
	* COMMENT AND DOCUMENT YOUR CODE, IDIOT! STOP MAKING FEATURES
	* Start thinking of ways to archive older applications that
	have either been successful or have been declined, or after a
	certain date would be classified as a "long-wait application".
	* Rosemary seems to only work well in vegtables. Makes meats
	too bland, like adding jasmine to your tea with nothing else.
	Try rosemary and roasted eggplant.

Good work, start getting ready for the creation of the simple web-page
form that will act as a server on the home computer so you no longer
have to create this via command-line anymore. This is great, but filling
out a webpage form will be a lot easier than popping open this specific
computer to do it each time.